4/10 4:42am
Forgor to mention it last time but fixed the collision error. Added animation trigger for upon pickup, along with transition to next scene. Added animations to end screen as well. Some buttons are non clickable with mouse buts thats a-ok for current build. What is remaining is collision bounds for post-change, timer for joystick, and sound!